refactor: 회원가입 스탭 atom 네이밍 수정
