--- src/recoil/auth/signUp/index.ts
@@ -1,3 +1,0 @@
-import signUpFormData from "./atom";
-
-export default signUpFormData;
