+++ src/recoil/auth/signUpStep/index.ts
@@ -1,0 +1,3 @@
+import signUpStepAtom from './atom';
+
+export default signUpStepAtom;
